gh-634 Refactor XmlDataWriters to extract abstract builder
