+++ stroom-util/src/test/java/stroom/util/test/data/XmlAttributesDataWriterBuilder.java
@@ -1,10 +1,7 @@
 package stroom.util.test.data;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * Produces XML like:
@@ -25,11 +22,8 @@ import java.util.stream.Stream;
  *   </records>
  * </pre>
  */
-public class XmlAttributesDataWriterBuilder {
+public class XmlAttributesDataWriterBuilder extends AbstractXmlDataWriterBuilder {
 
-    private Optional<String> namespace = Optional.empty();
-    private String rootElementName = "records";
-    private String recordElementName = "record";
     private String fieldValueElementName = "data";
 
     public static XmlAttributesDataWriterBuilder builder() {
@@ -41,42 +35,31 @@ public class XmlAttributesDataWriterBuilder {
                 .build();
     }
 
+    @Override
     public XmlAttributesDataWriterBuilder namespace(final String namespace) {
-        this.namespace = Optional.of(namespace);
+        super.namespace(namespace);
         return this;
     }
 
+    @Override
     public XmlAttributesDataWriterBuilder rootElementName(final String rootElementName) {
-        this.rootElementName = rootElementName;
+        super.rootElementName(rootElementName);
         return this;
     }
 
+    @Override
     public XmlAttributesDataWriterBuilder recordElementName(final String recordElementName) {
-        this.recordElementName = recordElementName;
+        super.recordElementName(recordElementName);
         return this;
     }
 
-    public XmlAttributesDataWriterBuilder fieldValueElementName(final String fielValueElementName) {
-        this.fieldValueElementName = fielValueElementName;
+    public XmlAttributesDataWriterBuilder fieldValueElementName(final String fieldValueElementName) {
+        this.fieldValueElementName = fieldValueElementName;
         return this;
     }
 
-    public DataWriter build() {
-        //return our mapping function which conforms to the DataWriter interface
-        return this::mapRecords;
-    }
-
-    private Function<Record, String> getDataMapper(final List<Field> fields) {
-        final String recordFormatStr = buildRecordFormatString(fields);
-
-        return record -> {
-            String[] valuesArr = new String[record.getValues().size()];
-            record.getValues().toArray(valuesArr);
-            return String.format(recordFormatStr, (Object[]) valuesArr);
-        };
-    }
-
-    private final String buildRecordFormatString(final List<Field> fields) {
+    @Override
+    protected final String buildRecordFormatString(final List<Field> fields) {
         final String fieldsPart = fields.stream()
                 .map(field -> "<" + fieldValueElementName + " name=\"" + field.getName() + "\" value=\"%s\" />")
                 .collect(Collectors.joining());
@@ -86,20 +69,4 @@ public class XmlAttributesDataWriterBuilder {
                 + "</" + recordElementName + ">";
     }
 
-    private Stream<String> mapRecords(List<Field> fields, Stream<Record> recordStream) {
-        final Function<Record, String> dataMapper = getDataMapper(fields);
-
-        final Stream<String> dataStream = recordStream.map(dataMapper);
-        final String namespaceAtr = namespace
-                .map(namespace -> String.format(" xmlns=\"%s\"", namespace))
-                .orElse("");
-
-        final String xmlDeclaration = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>";
-        final String openRootElm = String.format("<%s%s>", rootElementName, namespaceAtr);
-        final String closeRootElm = String.format("</%s>", rootElementName);
-
-        final Stream<String> headerStream = Stream.of(xmlDeclaration, openRootElm);
-        final Stream<String> footerStream = Stream.of(closeRootElm);
-        return Stream.concat(Stream.concat(headerStream, dataStream), footerStream);
-    }
 }
